+++ Tic-Tac-Toe with AI/task/src/tictactoe/Main.java
@@ -2,7 +2,8 @@ package tictactoe;
 
 import tictactoe.opponents.*;
 
-import static tictactoe.utils.GameStatus.*;
+import static tictactoe.utils.GameStatus.hasFreeSpace;
+import static tictactoe.utils.UserInteraction.getUserInput;
 import static tictactoe.utils.TicTacGraphics.*;
 
 public class Main {
@@ -31,8 +32,8 @@ public class Main {
         }
 
         try {
-            firstOpponent = LEVEL.valueOf(params[1].toUpperCase()).createOpponent();
-            secondOpponent = LEVEL.valueOf(params[2].toUpperCase()).createOpponent();
+            firstOpponent = Level.valueOf(params[1].toUpperCase()).createOpponent();
+            secondOpponent = Level.valueOf(params[2].toUpperCase()).createOpponent();
         } catch (Exception exc) {
             System.out.println("Bad parameters!");
             return false;
@@ -52,9 +53,9 @@ public class Main {
     static void play(GameChar[][] board) {
         while (true) {
             while (!firstOpponent.makeMove(board)) {}
-            if (!isPlace(board)) break;
+            if (!hasFreeSpace(board)) break;
             while (!secondOpponent.makeMove(board)) {}
-            if (!isPlace(board)) break;
+            if (!hasFreeSpace(board)) break;
         }
     }
 }
