+++ Tic-Tac-Toe with AI/task/src/tictactoe/opponents/Player.java
@@ -4,6 +4,7 @@ import tictactoe.GameChar;
 
 import static tictactoe.utils.GameStatus.*;
 import static tictactoe.utils.TicTacGraphics.*;
+import static tictactoe.utils.UserInteraction.getUserInput;
 
 public class Player extends Opponent {
     @Override
